COMP: move `await` completion from keyword completion contributor
`await` completion requires type inference, i.e. may uses indices under the hood. But keyword completion is supposed to be launched in dumb mode and can't uses indices because it leads to IndexNotReadyException
