+++ src/test/kotlin/org/rust/lang/core/completion/RsCompletionTestBase.kt
@@ -72,6 +72,12 @@ abstract class RsCompletionTestBase : RsTestBase() {
         @Language("Rust") code: String
     ) = completionFixture.checkContainsCompletion(code, variant)
 
+    protected fun checkCompletion(
+        lookupString: String,
+        @Language("Rust") before: String,
+        @Language("Rust") after: String
+    ) = completionFixture.checkCompletion(lookupString, before, after)
+
     protected fun checkNotContainsCompletion(
         variant: String,
         @Language("Rust") code: String
